+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## Version 0.0.4
+* support apply command to the select file/directory on explorer by use context menu item `SVN ... (Select Action) `
+* add new commands for current focusing file
+* directory remove `blame` command
+* modify command name and command title
+* remove menu `tortoise-svn log in workspace` to `explorer/context` and `editor/context` 
+* improve keybindings configuration
+* some bug fixed
+
 ## Version 0.0.3
 * add command `tortoise-svn...` which can select TortoiseSVN action and target dir or file
 * add menu `tortoise-svn...` to `explorer/context` and `editor/context`